--- src/main/java/com/nova/yeoback/global/dto/request/RequestDTO.java
@@ -1,4 +1,0 @@
-package com.nova.yeoback.global.dto.request;
-
-public class RequestDTO {
-}
