--- src/main/java/com/nova/yeoback/global/dto/response/ResponseDTO.java
@@ -1,4 +1,0 @@
-package com.nova.yeoback.global.dto.response;
-
-public class ResponseDTO {
-}
