+++ src/main/java/com/nova/yeobaek/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.nova.yeoback.global.config;
+package com.nova.yeobaek.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
