+++ src/main/java/com/nova/yeobaek/global/dto/request/RequestDTO.java
@@ -1,0 +1,4 @@
+package com.nova.yeobaek.global.dto.request;
+
+public class RequestDTO {
+}
