+++ src/main/java/com/nova/yeobaek/global/dto/response/ResponseDTO.java
@@ -1,0 +1,4 @@
+package com.nova.yeobaek.global.dto.response;
+
+public class ResponseDTO {
+}
