+++ src/main/java/com/nova/yeobaek/global/error/handler/ExceptionAdvice.java
@@ -1,4 +1,4 @@
-package com.nova.yeoback.global.error.handler;
+package com.nova.yeobaek.global.error.handler;
 
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -20,10 +20,10 @@ import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
-import com.nova.yeoback.global.error.exception.GeneralException;
-import com.nova.yeoback.global.error.response.CommonResponse;
-import com.nova.yeoback.global.error.status.CommonErrorStatus;
-import com.nova.yeoback.global.error.status.ErrorReason;
+import com.nova.yeobaek.global.error.exception.GeneralException;
+import com.nova.yeobaek.global.error.response.CommonResponse;
+import com.nova.yeobaek.global.error.status.CommonErrorStatus;
+import com.nova.yeobaek.global.error.status.ErrorReason;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolation;
